featureDev: Move streaming client to shared client
Problem:
- Other folders were piggybacking off the streaming client

Solution:
- Move to it src/clients so it's in a more central place
