+++ packages/core/src/amazonqFeatureDev/client/featureDev.ts
@@ -3,9 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { CodeWhispererStreaming } from '@amzn/codewhisperer-streaming'
 import { Service, Token } from 'aws-sdk'
-import { ConfiguredRetryStrategy } from '@aws-sdk/util-retry'
 import { omit } from 'lodash'
 import { AuthUtil } from '../../codewhisperer/util/authUtil'
 import { ServiceOptions } from '../../shared/awsClientBuilder'
@@ -19,6 +17,7 @@ import { ApiError, ContentLengthError, UnknownApiError } from '../errors'
 import { ToolkitError, isAwsError, isCodeWhispererStreamingServiceException } from '../../shared/errors'
 import { getCodewhispererConfig } from '../../codewhisperer/client/codewhisperer'
 import { LLMResponseType } from '../types'
+import { createCodeWhispererChatStreamingClient } from '../../shared/clients/codewhispererChatClient'
 
 // Create a client for featureDev proxy client based off of aws sdk v2
 export async function createFeatureDevProxyClient(): Promise<FeatureDevProxyClient> {
@@ -42,21 +41,6 @@ export async function createFeatureDevProxyClient(): Promise<FeatureDevProxyClie
     )) as FeatureDevProxyClient
 }
 
-// Create a client for featureDev streaming based off of aws sdk v3
-async function createFeatureDevStreamingClient(): Promise<CodeWhispererStreaming> {
-    const bearerToken = await AuthUtil.instance.getBearerToken()
-    const cwsprConfig = getCodewhispererConfig()
-    const streamingClient = new CodeWhispererStreaming({
-        region: cwsprConfig.region,
-        endpoint: cwsprConfig.endpoint,
-        token: { token: bearerToken },
-        // SETTING max attempts to 0 FOR BETA. RE-ENABLE FOR RE-INVENT
-        // Implement exponential back off starting with a base of 500ms (500 + attempt^10)
-        retryStrategy: new ConfiguredRetryStrategy(0, (attempt: number) => 500 + attempt ** 10),
-    })
-    return streamingClient
-}
-
 const streamResponseErrors: Record<string, number> = {
     ValidationException: 400,
     AccessDeniedException: 403,
@@ -71,12 +55,6 @@ export class FeatureDevClient {
         // Should not be stored for the whole session.
         // Client has to be reinitialized for each request so we always have a fresh bearerToken
         return await createFeatureDevProxyClient()
-    }
-
-    public async getStreamingClient() {
-        // Should not be stored for the whole session.
-        // Client has to be reinitialized for each request so we always have a fresh bearerToken
-        return await createFeatureDevStreamingClient()
     }
 
     public async createConversation() {
@@ -145,7 +123,7 @@ export class FeatureDevClient {
         { responseType: 'EMPTY'; approach?: string } | { responseType: 'VALID' | 'INVALID_STATE'; approach: string }
     > {
         try {
-            const streamingClient = await this.getStreamingClient()
+            const streamingClient = await createCodeWhispererChatStreamingClient()
             const params = {
                 workspaceState: {
                     programmingLanguage: { languageName: 'javascript' },
@@ -255,7 +233,7 @@ export class FeatureDevClient {
 
     public async exportResultArchive(conversationId: string) {
         try {
-            const streamingClient = await this.getStreamingClient()
+            const streamingClient = await createCodeWhispererChatStreamingClient()
             const params = {
                 exportId: conversationId,
                 exportIntent: 'TASK_ASSIST',
