+++ packages/core/src/codewhisperer/service/transformationResultsViewProvider.ts
@@ -12,7 +12,6 @@ import vscode from 'vscode'
 
 import { ExportIntent } from '@amzn/codewhisperer-streaming'
 import { TransformByQReviewStatus, transformByQState } from '../models/model'
-import { FeatureDevClient } from '../../amazonqFeatureDev/client/featureDev'
 import { ExportResultArchiveStructure, downloadExportResultArchive } from '../../shared/utilities/download'
 import { ToolkitError } from '../../shared/errors'
 import { getLogger } from '../../shared/logger'
@@ -21,6 +20,7 @@ import { codeTransformTelemetryState } from '../../amazonqGumby/telemetry/codeTr
 import { calculateTotalLatency } from '../../amazonqGumby/telemetry/codeTransformTelemetry'
 import { MetadataResult } from '../../shared/telemetry/telemetryClient'
 import * as CodeWhispererConstants from '../models/constants'
+import { createCodeWhispererChatStreamingClient } from '../../shared/clients/codewhispererChatClient'
 
 export abstract class ProposedChangeNode {
     abstract readonly resourcePath: string
@@ -257,10 +257,8 @@ export class ProposedTransformationExplorer {
     private changeViewer: vscode.TreeView<ProposedChangeNode>
 
     public static TmpDir = os.tmpdir()
-    private featureDevClient
 
     constructor(context: vscode.ExtensionContext) {
-        this.featureDevClient = new FeatureDevClient()
         const diffModel = new DiffModel()
         const transformDataProvider = new TransformationResultsProvider(diffModel)
         this.changeViewer = vscode.window.createTreeView(TransformationResultsProvider.viewType, {
@@ -309,7 +307,7 @@ export class ProposedTransformationExplorer {
                 transformByQState.getJobId(),
                 'ExportResultsArchive.zip'
             )
-            const cwStreamingClient = await this.featureDevClient.getStreamingClient()
+            const cwStreamingClient = await createCodeWhispererChatStreamingClient()
             try {
                 await downloadExportResultArchive(
                     cwStreamingClient,
